+++ src/main/kotlin/Hero.kt
@@ -1,13 +1,13 @@
 import kotlin.math.min
 
-abstract class Hero(val name: String, val maxHp: Int) {
+abstract class Hero(val name: String, val maxHp: Double) {
     var hp = maxHp
     var skillMod = 1.0
     var tenacity = 1.0
     var cantHeal = false
     var cantHealTimer = 0
 
-    fun heal(healAmount: Int) {
+    fun heal(healAmount: Double) {
         hp = min(hp + healAmount, maxHp)
     }
